Refactor stats modal
