+++ src/features/images/statsComponents/StatsCard.jsx
@@ -1,0 +1,14 @@
+import { styled } from "@stitches/react";
+
+const StatsCard = styled('div', {
+  background: '$loContrast',
+  padding: '15px',
+  border: '1px solid $border',
+  borderRadius: '5px',
+  width: '100%',
+  height: '100%',
+  overflowX: 'scroll',
+  overflowY: 'scroll',
+});
+
+export default StatsCard;
